+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageIdBuilder.swift
@@ -16,6 +16,9 @@ public class DefaultMessageIdBuilder: MessageIdBuilder {
     public init() { /* Public init. */ }
 
     public func makeMessageId(for message: ChatMessage) -> String {
+        if StreamRuntimeCheck._isDatabaseObserverItemReusingEnabled {
+            return message.id
+        }
         var statesId = "empty"
         if message.localState != nil {
             statesId = message.uploadingStatesId