+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageActions_Tests.swift
@@ -72,6 +72,44 @@ class MessageActions_Tests: StreamChatTestCase {
         XCTAssert(messageActions[5].title == "Flag Message")
         XCTAssert(messageActions[6].title == "Mute User")
     }
+    
+    func test_messageActions_otherUserDefaultBlockingEnabled() {
+        // Given
+        streamChat = StreamChat(
+            chatClient: chatClient,
+            utils: .init(messageListConfig: .init(userBlockingEnabled: true))
+        )
+        let channel = ChatChannel.mockDMChannel()
+        let message = ChatMessage.mock(
+            id: .unique,
+            cid: channel.cid,
+            text: "Test",
+            author: .mock(id: .unique),
+            isSentByCurrentUser: false
+        )
+        let factory = DefaultViewFactory.shared
+
+        // When
+        let messageActions = MessageAction.defaultActions(
+            factory: factory,
+            for: message,
+            channel: channel,
+            chatClient: chatClient,
+            onFinish: { _ in },
+            onError: { _ in }
+        )
+
+        // Then
+        XCTAssert(messageActions.count == 8)
+        XCTAssert(messageActions[0].title == "Reply")
+        XCTAssert(messageActions[1].title == "Thread Reply")
+        XCTAssert(messageActions[2].title == "Pin to conversation")
+        XCTAssert(messageActions[3].title == "Copy Message")
+        XCTAssert(messageActions[4].title == "Mark Unread")
+        XCTAssert(messageActions[5].title == "Flag Message")
+        XCTAssert(messageActions[6].title == "Mute User")
+        XCTAssert(messageActions[7].title == "Block User")
+    }
 
     func test_messageActions_currentUserPinned() {
         // Given